patch update : late

+++ resources/views/attendances/employee_attendance.blade.php
@@ -270,8 +270,8 @@
                                             @if(in_array(date('l',strtotime($date_r)),$schedules->pluck('name')->toArray()) && $time_in_data)
                                                 @php
                                                   $id = array_search(date('l',strtotime($date_r)),$schedules->pluck('name')->toArray());
-                                                  $late =  (double) (strtotime(date("01-01-2022 h:i",strtotime($time_in_data))) - strtotime(date("01-01-2022 h:i",strtotime("01-01-2022 ".$schedules[$id]->time_in_to))))/60;
-                                                  $working_minutes = (double) (((strtotime($time_in->time_out) - strtotime($time_in_data)))/3600);
+                                                  $late =  (double) (strtotime(date("01-01-2022 h:i",strtotime($time_in_data))) - (double) strtotime(date("01-01-2022 h:i",strtotime("01-01-2022 ".$schedules[$id]->time_in_to))))/60;
+                                                  $working_minutes = (double) (((strtotime($time_in->time_out) - (double) strtotime($time_in_data)))/3600);
                                                   $overtime = (double) number_format($working_minutes - $schedules[$id]->working_hours,2);
                                                   if($late > 0)
                                                   {
